requirements.txt
fastapi
uvicorn
torch
torch-geometric
scikit-learn
joblib
numpy
matplotlib
seaborn
pymongo
python-dotenv